fix(opencode): extract task-context budget helpers under module-size ceiling
Module Size Guard and substrate hotspot ratchet failed after inlining the
BI-B195F224 trim path into opencode-dispatch (1043 LOC hard-cap / +1 hotspot).

Move pure estimate/budget/trim helpers to opencode-task-context-budget.ts so
dispatch stays at 799 LOC (under the 800 soft ceiling) with no hotspot growth.

Process-Spine-Decision: pure bug fix with regression tests; no new durable surface
